pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "TaskfileToTasks"
version = "1.0.0"
description = "Convert Taskfile.yml to tasks.json for VSCode or Zed editors"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "H3mul", email = "h3mul@github.com"}
]
keywords = ["taskfile", "tasks", "vscode", "zed", "task-runner"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Utilities",
]
requires-python = ">=3.8"
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "types-PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/H3mul/TaskfileToTasks"
Repository = "https://github.com/H3mul/TaskfileToTasks.git"
Documentation = "https://github.com/H3mul/TaskfileToTasks#readme"
"Bug Tracker" = "https://github.com/H3mul/TaskfileToTasks/issues"

[project.scripts]
TaskfileToTasks = "taskfile_to_tasks.cli:main"

[tool.setuptools]
packages = ["taskfile_to_tasks"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true

[[tool.mypy.overrides]]
module = "yaml.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers -v --tb=short"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
]